extensions-contrib/druid-deltalake-extensions/src/test/resources/requirements.txt
delta-spark==2.2.0
mimesis==5.0.0
pyspark==3.2.0